Wrap all functionality in a class
