+++ pyproject.toml
@@ -15,8 +15,9 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Topic :: Home Automation",
     "Topic :: Utilities",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.8"
